+++ src/axom/primal/operators/detail/intersect_ray_impl.hpp
@@ -29,7 +29,8 @@ namespace detail
  * When there is a valid intersection (within tolerance \a EPS), 
  * \a ray_param returns the parametric coordinate of the intersection point along \a R
  * and \a seg_param returns the parametric coordinate of the intersection point along \a S.
- *
+ * If the intersection point is nonunique, \a seg_param and \a ray_param return only
+ * a single point of intersection
  * \return status true iff R intersects with S, otherwise, false.
  */
 template <typename T>
@@ -41,23 +42,45 @@ inline bool intersect_ray(const primal::Ray<T, 2>& R,
 {
   AXOM_STATIC_ASSERT(std::is_floating_point<T>::value);
 
-  // STEP 0: Find the parameterized direction of the segment
+  // Find the parameterized direction of the segment.
   const auto seg_dir = primal::Vector<T, 2>(S.source(), S.target());
   const auto& ray_dir = R.direction();
 
-  // Step 1: Equating R(ray_param)=S(seg_param) yields a system of two equations and
+  // Define lower/upper threshold
+  const double tlow = 0.0 - EPS;
+  const double thigh = 1.0 + EPS;
+
+  // Equating R(ray_param)=S(seg_param) yields a system of two equations and
   // two unknowns, namely, t0 and t1. We can solve this system directly using Cramer's Rule.
   const double denom =
     numerics::determinant(ray_dir[0], -seg_dir[0], ray_dir[1], -seg_dir[1]);
 
-  // STEP 2: if denom is (nearly) zero (within tolerance EPS), the system is singular
+  // If denom is (nearly) zero (within tolerance EPS), the ray and segment are parallel
   if(axom::utilities::isNearlyEqual(denom, 0.0, EPS))
   {
-    // ray and segment are parallel
-    return false;
+    // Check if ray and segment are collinear
+    const auto col = S.source().array() - R.origin().array();
+
+    const double cross =
+      numerics::determinant(col[0], ray_dir[0], col[1], ray_dir[1]);
+
+    if(axom::utilities::isNearlyEqual(cross, 0.0, EPS))
+    {
+      // Check orientation of segment relative to ray
+      const double t0 = col[0] * ray_dir[0] + col[1] * ray_dir[1];
+      const double t1 = t0 + seg_dir.dot(ray_dir);
+
+      // Assign (nonunique) parameters
+      ray_param = (t1 > t0) ? t1 : t0;
+      seg_param = (t1 > t0) ? 1.0 : 0.0;
+
+      return ((ray_param >= tlow) && (seg_param >= tlow) && (seg_param <= thigh));
+    }
+    else  // Not collinear, no intersection
+      return false;
   }
 
-  // STEP 3: Solve for the ray_param and seg_param directly using cramer's rule
+  // Solve for the ray_param and seg_param directly using Cramer's rule
   const auto sol = S.source().array() - R.origin().array();
 
   // Note: ray_param is an OUT parameter of this function
@@ -68,11 +91,7 @@ inline bool intersect_ray(const primal::Ray<T, 2>& R,
   seg_param =
     numerics::determinant(ray_dir[0], sol[0], ray_dir[1], sol[1]) / denom;
 
-  // STEP 4: Define lower/upper threshold
-  const double tlow = 0.0 - EPS;
-  const double thigh = 1.0 + EPS;
-
-  // STEP 5: Necessary and sufficient criteria for an intersection between
+  // Necessary and sufficient criteria for an intersection between
   // ray, R(t0),  and a finite segment S(t1) are:
   // 1. ray_param >= tlow w.r.t. the ray R(ray_param).
   // 2. tlow >= seg_param >= thigh w.r.t. the segment S(seg_param).